ia: shift to id-based search
Previously, mismatched ID and title will cause an error to
be thrown. This is undesired behavior, as the ID is never
actually visible on the user side. New behavior will use
exclusively the page ID (when available) and fall back to
using the title when not. Protection has also been added
to avoid adding a reply to a listing that was detected as
malformed but is not actually malformed, which is an
indication of failed search.